requirements.txt
absl-py==0.7.1
astor==0.8.0
astroid==2.2.5
backcall==0.1.0
bidict==0.18.3
biosppy==0.6.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
entrypoints==0.3
flake8==3.7.7
future==0.17.1
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.27.2
h5py==2.9.0
heartpy==1.2.4
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.0
joblib==0.13.2
jupyter-client==5.2.4
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
matplotlib==3.1.0
mccabe==0.6.1
numpy==1.16.0
oauthlib==3.1.0
pandas==0.24.2
parso==0.5.0
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pkg-resources==0.0.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.3
pyzmq==18.0.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.2
scipy==1.3.0
shortuuid==0.5.0
six==1.12.0
tensorboard==2.1.0
tensorboardX==1.6
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.2.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.8
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.2
